chore(backend): Update validations
- Update redFlag validations to return the right response on error
- Fix travis build
- Rename validator to follow file structures

[Fixes #162439563]
